requirements.txt
pillow>=10.0.0,<11.0.0
numpy>=1.24.0,<2.0.0
requests>=2.31.0,<3.0.0
python-escpos>=3.0.0,<4.0.0
pyusb>=1.2.1,<2.0.0
